fix follow up issues
