+++ mock-platform/config/task-binary-map.json
@@ -435,10 +435,6 @@
         {
           "src": "tasks/social-cross-publish/environment/seed.json",
           "dest": "/opt/mock/data/seed.json"
-        },
-        {
-          "src": "tasks/common/verify_utils.py",
-          "dest": "/workspace/environment/verify_utils.py"
         }
       ]
     },
