+++ tasks/social-event-campaign/task.toml
@@ -1,7 +1,7 @@
 version = "1.0"
 
 [metadata]
-difficulty = "medium"
+difficulty = "easy"
 category = "open-world"
 tags = ["social_media", "communication_email"]
