+++ tasks/social-schedule-audit/task.toml
@@ -1,7 +1,7 @@
 version = "1.0"
 
 [metadata]
-difficulty = "medium"
+difficulty = "easy"
 category = "open-world"
 tags = ["social_media"]
